Add a notebook
The notebook clones the code repository, repeats the steps described
in its readme to obtain raw data (slightly tweaked, e.g. symlink), and
runs the code to generate fig 3.

Cloning is quick, but execution takes about 15 minutes.

FMPOV, this proves that we can run (some) of the code in a notebook in
Binder.