Add datalad to the conda environment
We run this as root because the conda env in the base image was also
created by root, and user install may or may not fail there.

We use pip because conda install would probably perform more costly
dependency resolution. We can do this, because git-annex (old, but
probably fine) is already part of the base env.