Add a patch modifying code to output PDF
This way we can use nawrotlab/ClusteredNetwork_pub v1.0.0 as the
source of demo code, and still output PNG, without the need to apply
changes upstream. Applying a patch is a single line, so does not
interrupt the narrative too much, it is transparent (the patch file
can be inspected), and it can be extended with further changes in the
future.

We prefer PNG over PDF because it can be shown in the bash kernel
notebook with `cat file.png | display` (custom function from the
jupyter kernel). Additionally, we have found PNG to be byte-exact
reproducible, however our demo will not show that.