Set git's advice.detachedHead to false in Dockerfile
Since in our demo notebook we specifying a tag upon cloning, setting
advice.detachedHead to false will disable showing the admonition about
detached head. While generally useful, it takes too much space on the
screen when walking through the demo, and it is unlikely that anyone
will want to interact with branches as part of their exploration
anyway.